Udp updates. Removed a fail-fast which could occur in the UDP path if the server were overloaded with client requests and a timer fired during closing. Removed the remaining references to "retried" UDP packets - that protocol property had already been removed a long while back. Updated the UDP output to show # of connections now that we don't need a column for "retries".

+++ ctsTraffic/ctsMediaStreamServerConnectedSocket.cpp
@@ -177,7 +177,7 @@ namespace ctsTraffic {
                 default:
                     ctl::ctAlwaysFatalCondition(
                         L"Unexpected task action returned from initiate_io - %u (dt %p ctsTraffic::ctsIOTask)",
-                        static_cast<unsigned int>(current_task.ioAction),
+                        static_cast<unsigned long>(current_task.ioAction),
                         &current_task);
             }
         }
